new_backend_ruminate/requirements.txt
fastapi==0.108.0
uvicorn==0.34.0
pydantic==2.10.6
sqlalchemy==2.0.23
alembic==1.14.0
asyncpg==0.30.0
pytest==7.4.3
pytest-asyncio==0.21.0
starlette<0.33.0,>=0.29.0
sse-starlette==1.8.2
anyio==4.8.0
aiofiles==24.1.0
aiohttp==3.11.2
async-timeout==5.0.1
click==8.1.7
email_validator==2.2.0
httpx==0.24.1
httpx-sse==0.4.0
Jinja2==3.1.4
openai==0.28.0
python-dotenv==1.0.0
python-multipart==0.0.9
pyyaml==6.0.1
setuptools>=60.0.0
typing_extensions==4.12.2
websockets==13.1
attrs==24.2.0
annotated-types==0.7.0
pydantic-settings==2.6.1
certifi>=2023.11.17
aioboto3==13.3.0
PyJWT==2.8.0
aiohttp==3.11.2
